requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
distro==1.9.0
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.13.2
fastapi==0.128.6
filelock==3.20.3
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
preshed==3.0.12
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.4.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_openai==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
wrapt==2.1.1
